requirements.txt
pandas
pygments
markdown
django-filter
djangorestframework
django-mptt
django-modeltranslation>=0.9
django-tables2
crispy-forms-foundation
django-crispy-forms
feedparser
django-reversion
Pillow
django-filebrowser
django-admin-tools
django-grappelli>=2.6
# django-livereload
# livereload
-e git://github.com/ojii/django-sekizai.git#egg=sekizai
django-jquery
django-bootstrap3
django_compressor
django-debug-toolbar
dj-database-url
django-toolbelt
django==1.8
alabaster==0.7.3
Babel==1.3
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
docutils==0.12
gnureadline==6.3.3
ipdb==0.8
ipython==3.1.0
Jinja2==2.7.3
jsonschema==2.4.0
MarkupSafe==0.23
mistune==0.5.1
nose==1.3.6
numpydoc==0.5
pep8==1.6.2
ptyprocess==0.4
pudb==2015.2
pyflakes==0.8.1
Pygments==2.0.2
pytz==2015.2
pyzmq==14.5.0
requests==2.6.0
see==1.1.0
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
terminado==0.5
tornado==4.1
urwid==1.3.0
